LogLevelByNameConfig - wip
